# Absorb some instance analysis into InstanceGraph, use safer boxed Strings