@jackkoenig @albertchen-sifive 

This absorbs some repetitive instance hierarchy analysis from `InlineInstances` and `ConstantPropagation` into `InstanceGraph`. In this version, strings are hidden with boxed types that have some potential to help with type safety when maintaining passes.

`InlineInstances` already uses the `TargetToken`s to hide strings, so I just went through and ported `ConstantPropagation` to get rid of raw `String` as much as possible. It does get rid of some scarier type signatures like `(NodeMap, Map[String, String], Map[String, Map[String, Literal]]) => ...`. I added some implicit classes to make it possible to write things like `inst.OfModule` or `inst.Instance`, which made the code a bit less cluttered in the few places where conversions happen.

Absorbing the instance analysis into API calls reduces both complexity (for example, `staticInstanceCount` is more efficient than existing code) and code size a bit, but this is offset by the overhead of boxing. I could alternatively add these APIs as versions that work with raw `String`s, which might be faster in theory, though I will be testing if that is the case in practice. This PR has no measurable performance impact on the integration test benchmark performance, but I am still testing bigger designs.

The new code in `InstanceGraph` returns collections with stable iteration orders in all cases.