@jackkoenig 2019-09-21
From my benchmarking, this looks like a nice speedup for small designs and a wash to perhaps minor speedup for small designs. I'm a little surprised by the size of the speedup for the small designs so I'm rerunning the benchmarks.

Design | master runtime (SD) | this runtime (SD) | this/master (avg) | this/master (worst) | this/master (best)
-- | -- | -- | -- | -- | --
FPU | 13.71 (3.56) | 4.6 (0.27) | 0.34 | 0.48 | 0.25
HwachaSequencer | 39.36 (10.74) | 13.02 (0.75) | 0.33 | 0.48 | 0.24
Icache | 5.68 (0.99) | 2.54 (0.1) | 0.45 | 0.56 | 0.37
Rob | 52.24 (14) | 29.45 (2.56) | 0.56 | 0.84 | 0.41
RocketCore | 11.05 (4.44) | 5.09 (0.38) | 0.46 | 0.83 | 0.30
SiFive A | 328.58 (65.32) | 329.5 (38) | 1.00 | 1.40 | 0.74
SiFive B | 80.75 (2.81) | 78.53 (2.51) | 0.97 | 1.04 | 0.91
SiFive C | 219.78 (13.15) | 219.26 (11.2) | 1.00 | 1.12 | 0.89
SiFive D | 90.75 (3.09) | 83.64 (1.58) | 0.92 | 0.97 | 0.87

Notes:
* "avg" - simply dividing the runtimes
* "worst" - (this + SD) / (master - SD)
* "best" - (this - SD) / (master + SD)