@albert-magyar 2019-09-23
That is a little .. too good. I was seeing a speedup in some smaller designs, but 3x seems kind of too high.