@jackkoenig 2019-09-23
Updated after rerunning the smaller configs in `regress/`. A bit more reasonable and the high variance makes it obvious not to read too much into the numbers.


Design | master runtime (SD) | this runtime (SD) | this/master (avg) | this/master (worst) | this/master (best)
-- | -- | -- | -- | -- | --
FPU | 6.98 (3.66) | 4.28 (0.28) | 0.61 | 1.37 | 0.38
HwachaSequencer | 16.73 (5.49) | 11.78 (0.6) | 0.70 | 1.10 | 0.50
Icache | 3.74 (1.4) | 2.17 (0.15) | 0.58 | 0.99 | 0.39
Rob | 27.39 (11.52) | 25.81 (2.54) | 0.94 | 1.79 | 0.60
RocketCore | 4.78 (0.64) | 4.84 (0.24) | 1.01 | 1.23 | 0.85
SiFive A | 328.58 (65.32) | 329.5 (38) | 1.00 | 1.40 | 0.74
SiFive B | 80.75 (2.81) | 78.53 (2.51) | 0.97 | 1.04 | 0.91
SiFive C | 219.78 (13.15) | 219.26 (11.2) | 1.00 | 1.12 | 0.89
SiFive D | 90.75 (3.09) | 83.64 (1.58) | 0.92 | 0.97 | 0.87