@jackkoenig 2019-09-23
It's weird that this PR makes the standard deviation so much lower