@albert-magyar 2019-09-23
It's also funny, since the point of the changes were to make it "safer" by relying more on the type system to distinguish types of strings without slowing things down, but if the `InstanceGraph` API calls are a bit faster that's good :)