@albert-magyar 2019-09-29
I can take them out, it's not a huge deal.

I figure the real risk would be actually having an implicit conversion `T => OfModule`. This still requires you to call a helper method of the implicit class to convert, so it's basically just syntactic sugar compared to making it `OfModule.apply(Module)` or something.